% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{data_std}
\alias{data_std}
\title{A matrix of counts
A dataset of gene expression profile, a large matrix with 21 rows and 2603
columns.}
\format{
Gene list obtained by differential gene analysis
}
\description{
A matrix of counts
A dataset of gene expression profile, a large matrix with 21 rows and 2603
columns.
}
\examples{
data(data_std)
}
