% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{gene_list_hsa}
\alias{gene_list_hsa}
\title{A gene list of human}
\format{
Gene list obtained by differential gene analysis
}
\description{
A dataset of gene list
}
\examples{
# Load the data.
data(gene_list_hsa)
}
