% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{get_bg_related_kegg}
\alias{get_bg_related_kegg}
\title{get_bg_related_kegg}
\usage{
get_bg_related_kegg(gene_list, PFPRefnet, rm_duplicated = FALSE)
}
\arguments{
\item{gene_list, }{a vector of characters, refers to genes ids}

\item{PFPRefnet, }{an object of PFPRefnet class, it contains all kegg pathways.}

\item{rm_duplicated, }{a logical, whether to remove the duplicated kegg edges in different pathways.
Defalut is \emph{FALSE}}
}
\value{
the related kegg network.
}
\description{
This function will select all genes in all kegg pathways which are directly connected with the genes in \code{gene_list}
}
\details{
, It will return a data.frame which can be translated a graph or network.
In the data.frame, \code{source} refers to the genes in \code{gene_list}, \code{target} refers to the directly
connected genes in kegg, \code{weight} is 0.5, no real means, \code{pathway}
refers to the pathway which the edge emerge and \code{edge_type} is "kegg".
Note, if \code{rm_duplicated} is \emph{FALSE}, it may return many duplicated edges,
which will be complex when plotting a network. If \code{rm_duplicated} is \emph{TRUE},
it will retain the first pathway which contains the duplicated edge.
}
\examples{
# Load the PFPRefnet of human
data(PFPRefnet_hsa)
# Load the list of diff genes
data(gene_list)
# Get the related kegg network
edges_kegg <- get_bg_related_kegg(gene_list,PFPRefnet_hsa)
}
