% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kegg_graphs.R
\name{get_pathway_info}
\alias{get_pathway_info}
\title{get pathway info of a species in KEGG}
\usage{
get_pathway_info(spec)
}
\arguments{
\item{spec, }{a character, refers to the species in KEGG. hsa, mmu...}
}
\value{
a data.frame whose colnames contains "index","id","name" and "group"
}
\description{
This function helps get pathway info of a species in KEGG.
}
\details{
, get pathway info of a species in KEGG. It will
return a data.frame.
}
\examples{
pathway_info <- get_pathway_info("hsa")
}
