% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PFP_score.R
\name{get_PFP_score}
\alias{get_PFP_score}
\title{Get the pathway fingerprint of a gene_list}
\usage{
get_PFP_score(
  genes,
  PFPRefnet,
  coeff1 = 1,
  coeff2 = 0.1,
  statistic = FALSE,
  bg_genelist = NULL,
  num_loop = 10
)
}
\arguments{
\item{genes, }{a vector of characters}

\item{PFPRefnet, }{A PFPRefnet class}

\item{coeff1, }{a numeric, the coefficient 1 in PFP model}

\item{coeff2, }{a numeric, the coefficient 2 in PFP model}

\item{statistic, }{a logical,whether to do the statistical test}

\item{bg_genelist, }{a vector of characters, background gene set for the statistical test}

\item{num_loop, }{random sampling times from bg_genelist for the statistical test}
}
\description{
It can evaluate the performance of a gene list in the pathway networks.
}
\details{
The main part of pathway fingerprint. PFP is used to evaluate the performance of
a gene_list in some pathway networks by considering the genes' topological location in
a pathway. Then we can get every gene's score and the pathway score is caculated by sum
all genes' score. All pathways' scores combine the pathway fingerprint.
}
\examples{
\dontrun{
genes <- c("101488143","15129","15122","110257","14102")
load("PFPRefnet")
PFP <- get_PFP_score(genes,PFPRefnet)
}
}
