% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kegg_graphs.R
\name{get_PFPRefnet}
\alias{get_PFPRefnet}
\title{get a PFPRefnet for a species}
\usage{
get_PFPRefnet(spec, file_root = ".", test_mode = FALSE)
}
\arguments{
\item{spec, }{a character, refers to the species in KEGG. hsa, mmu...}

\item{file_root, }{a character, file dir to download the kgml files.}

\item{test_mode, }{please set whether to test this function.}
}
\value{
a PFPRefnet object.
}
\description{
This function helps update the latest PFPRefnet odject for a species
}
\details{
, gupdate the latest PFPRefnet odject for a species in KEGG. It will
return a PFPRefnet object.
}
\examples{
PFPRefnet1 <- get_PFPRefnet("hsa",".",test_mode=TRUE)
}
