% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{group-methods}
\alias{group-methods}
\alias{group}
\alias{group,PFPRefnet-method}
\title{group information of \emph{PFPRefnet}}
\usage{
group(object)

\S4method{group}{PFPRefnet}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} class}
}
\value{
a list contains names of basic groups of the networks and group number, as
well as the size of each group
}
\description{
This function contains names of basic groups of the networks and group number, as
well as the size of each group
}
\examples{
# New a PFPRefnet object
PFPRefnet1 <- new("PFPRefnet", network = list1, net_info = data1)
group <- group(PFPRefnet1)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
