% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{net_info-methods}
\alias{net_info-methods}
\alias{net_info}
\alias{net_info,PFPRefnet-method}
\title{Basic pathway networks information of \emph{PFPRefnet} class}
\usage{
net_info(object)

\S4method{net_info}{PFPRefnet}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} class}
}
\value{
a dataframe contains basic networks' information
}
\description{
This function extract the basic networks information of PFPRefnet class.
}
\examples{
# New a PFPRefnet object
data(PFPRefnet_hsa)
# PFPRefnet1 <- new("PFPRefnet", network = list1, net_info = data1)
net_info <- net_info(PFPRefnet_hsa)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
