% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{net_names-methods}
\alias{net_names-methods}
\alias{net_names}
\alias{net_names,PFPRefnet-method}
\title{Names of basic networks}
\usage{
net_names(object)

\S4method{net_names}{PFPRefnet}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} class}
}
\value{
a vector contains pathway names
}
\description{
This function extract the network names of PFPRefnet.
}
\examples{
# New a PFPRefnet object
PFPRefnet1 <- new("PFPRefnet", network = list1, net_info = data1)
net_names <- net_names(PFPRefnet1)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
