% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{network-methods}
\alias{network-methods}
\alias{network}
\alias{network,PFPRefnet-method}
\title{Basic pathway networks of \emph{PFPRefnet} class}
\usage{
network(object)

\S4method{network}{PFPRefnet}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} class}
}
\value{
a graphNEL list of all basic networks
}
\description{
This function extract the basic networks of PFPRefnet class.
}
\examples{
# New a PFPRefnet object
PFPRefnet1 <- new("PFPRefnet", network = list1, net_info = data1)
network <- network(PFPRefnet1)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
