% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{p_value-methods}
\alias{p_value-methods}
\alias{p_value}
\alias{p_value,PFP-method}
\title{The P value of \emph{PFP}
This function can extract the result of statistical analysis}
\usage{
p_value(object)

\S4method{p_value}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} class}
}
\value{
Statistical test result of each pathway score
}
\description{
The P value of \emph{PFP}
This function can extract the result of statistical analysis
}
\seealso{
\code{\link{PFP-class}}
}
