% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{pathway_info}
\alias{pathway_info}
\title{pathway_info
pathway_info is dataframe of the information of pathway, 539rows、4columns.}
\format{
a list
}
\description{
pathway_info
pathway_info is dataframe of the information of pathway, 539rows、4columns.
}
\examples{
data(pathway_info)
}
