% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{pathway_info_hsa}
\alias{pathway_info_hsa}
\title{pathway_info
pathway_info is dataframe of the information of pathway, 539rows, 4columns.}
\format{
a list
pathway_info_hsa
pathway_info_hsa is dataframe of the information of pathway, 539rows,
4columns. The main data is about human.
}
\description{
pathway_info
pathway_info is dataframe of the information of pathway, 539rows, 4columns.
}
\examples{
data(pathway_info_hsa)
}
