% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{class}
\name{PFP-class}
\alias{PFP-class}
\title{\code{PFP-class}}
\description{
An S4 object for storing pathway fingerprint scores information.
}
\section{Slots}{

\describe{
\item{\code{pathways_score,}}{a list contains PFP_score, p_value, random_score, genes_score.
PFP_score is a numeric score indicating the performance of a gene_list in some pathways.
p_value is a statistic test for the PFP_score.
random_score is used for the statistic test.
genes_score is the detail scores of every gene in the gene_list.}

\item{\code{ref_net_info,}}{a data.frame, which contains the specific information of pathway networks.
Just be the same as \code{\link{net_info}} in FPFRefnet, including the index, id, name, group and species.
#'@section method:
\itemize{
\item{pathways_score, \code{signature(object = "PFP")}:
extract the pathways score}
\item{ref_net_info, \code{signature(object = "PFP")}:
extract the pathway networks information}
\item{PFP_score, \code{signature(object = "PFP")}:
extract the PFP score}
\item{p_value, \code{signature(object = "PFP")}:
extract P value}
\item{random_score, \code{signature(object = "PFP")}:
extract the random score}
\item{genes_score, \code{signature(object = "PFP", index=NULL,
     index_type = c("pathway_id","pathway_name","slice"))}:
extract the genes score}
\item{refnet_names, \code{signature(object = "PFP")}:
extract the refnet names}
\item{sub_PFP, \code{signature(object = "PFP", group_name = NULL,
     index = NULL, index_type = c("slice","pathway_id","pathway_name"))}
subset of PFP object}
\item{plot_PFP, \code{signature(object, type = "character",
     p_size = "numeric", l_size = 'numeric')}:
plot the Pathway Fingerprint.}
\item{rank_PFP \code{signature(object = "PFP", decreasing=TRUE)}
sort the PFP score.}
\item{show_PFP, \code{signature(object = "PFP")}:
display methods for S4 classes PFP, see also
\code{\link[methods]{show_PFP}}}
}}
}}

\seealso{
\code{\link{pathways_score-methods}},
\code{\link{refnet_info-methods}}, \code{\link{PFP_score-methods}},
\code{\link{p_value-methods}}, \code{\link{random_score-methods}},
\code{\link{genes_score-methods}}, \code{\link{refnet_names-methods}},
\code{\link{sub_PFP-methods}}, \code{\link{plot_PFP-methods}},
\code{\link{rank_PFP-methods}}, \code{\link{show_PFP-methods}},
}
