% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{PFP_score-methods}
\alias{PFP_score-methods}
\alias{PFP_score}
\alias{PFP_score,PFP-method}
\title{The score of \emph{PFP}
This function can extract the PFP_score of PFP.}
\usage{
PFP_score(object)

\S4method{PFP_score}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} class}
}
\value{
the PFP_score
}
\description{
The score of \emph{PFP}
This function can extract the PFP_score of PFP.
}
\examples{
# New a PFP object
data(PFP_test1)
# s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
PFP_score <- PFP_score(PFP_test1)
}
\seealso{
\code{\link{PFP-class}}
}
