% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{PFP_test2}
\alias{PFP_test2}
\title{PFP_test2}
\format{
A PFPRefnet object,
}
\description{
A dataset of PFP class
}
\details{
a PFP class with pathways_score(a list of length 3), refnet_info
(a datafame, 338rows and 3 columns), the main data is about human.
}
\examples{
data(PFP_test2)
}
\seealso{
\code{\link{PFP-class}}
}
