% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{PFPRefnet_mmu}
\alias{PFPRefnet_mmu}
\title{Pathway fingerprint data of mouse}
\format{
A PFPRefnet object, more details see \code{\link{PFPRefnet-class}}
}
\description{
A dataset containing the pathway maps of KEGG
PFPRefnet_mmu  a PFPRefnet class with network(a list of length 334), net_info
(a datafame, 334rows and 5 columns), the main data is about mouse.
}
\examples{
data(PFPRefnet_mmu)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
