% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{plot_PFP-methods}
\alias{plot_PFP-methods}
\alias{plot_PFP}
\alias{plot_PFP,PFP-method}
\title{Plot PFP results
Function for visualization PFP results.}
\usage{
plot_PFP(
  object,
  type = c("matchstick", "line", "point"),
  p_size = 1,
  l_size = 0.5
)

\S4method{plot_PFP}{PFP}(
  object,
  type = c("matchstick", "line", "point"),
  p_size = 1,
  l_size = 0.5
)
}
\arguments{
\item{object, }{\code{PFP} class}

\item{type, }{types of the visaulization of \emph{PFP} object, 'matchstick', 'line','point'.
Default is 'matchstick'.}

\item{p_size, }{point size of plot, default is 1.}

\item{l_size, }{line size of plot, default is 0.5.}
}
\value{
a plot of PFP
}
\description{
Plot PFP results
Function for visualization PFP results.
}
\examples{
# New a PFP object
data(PFP_test1)
#s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
#plot_PFP(FPRefnet_hsa,'line', p_size = 1, l_size = 0.5)
}
\seealso{
\code{\link{PFP-class}}
}
