% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{random_score-methods}
\alias{random_score-methods}
\alias{random_score}
\alias{random_score,PFP-method}
\title{Extract the randomized similarity score of \emph{PFP}
This function extract the randomized similarity score for statistical test.}
\usage{
random_score(object)

\S4method{random_score}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} class}
}
\value{
a data.frame, each row represents one random test result with the same number of gene_list.
}
\description{
Extract the randomized similarity score of \emph{PFP}
This function extract the randomized similarity score for statistical test.
}
\seealso{
\code{\link{PFP}}
}
