% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{ref_net_info-methods}
\alias{ref_net_info-methods}
\alias{ref_net_info}
\alias{ref_net_info,PFP-method}
\title{Basic network information of \emph{PFP} class
This function extract the detail information of reference pathway networks.}
\usage{
ref_net_info(object)

\S4method{ref_net_info}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} class}
}
\value{
detail information of reference pathway networks
}
\description{
Basic network information of \emph{PFP} class
This function extract the detail information of reference pathway networks.
}
\seealso{
\code{\link{PFP-class}}
}
