% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{refnet_names-methods}
\alias{refnet_names-methods}
\alias{refnet_names}
\alias{refnet_names,PFP-method}
\title{Names of basic networks
This function extract the reference pathway network names of PFP.}
\usage{
refnet_names(object)

\S4method{refnet_names}{PFP}(object)
}
\arguments{
\item{object, }{\code{\link{PFPRefnet-class}} class}
}
\value{
a vector contains pathway names
}
\description{
Names of basic networks
This function extract the reference pathway network names of PFP.
}
