% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{result_PFP-methods}
\alias{result_PFP-methods}
\alias{result_PFP}
\alias{result_PFP,PFP-method}
\title{result of the PFP object.}
\usage{
result_PFP(object, thresh_slot = NULL, thresh_value = 0.05)

\S4method{result_PFP}{PFP}(object, thresh_slot = NULL, thresh_value = 0.05)
}
\arguments{
\item{object, }{\code{PFP} class}

\item{thresh_slot, }{a character, it could be 'p_value' or 'p_adj_value',
it means the threshold slot to choose for select the significant pathway.
Default is \emph{NULL},it means that you don't
want to select the significant pathway and you will select all pathways.}

\item{thresh_value, }{a numeric, threshold value of 'p_value' or
'p_adjust_value' for pathway selection,Default is 0.05.}
}
\value{
the scores and the information of PFP object.
}
\description{
get the result of the PFP object.
}
\examples{
data(PFP_test1)
result_PFP(PFP_test1,
        thresh_slot="p_adj_value",
        thresh_value = 0.05)
}
\seealso{
\code{\link{PFP-class}}
}
