% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{show_net-methods}
\alias{show_net-methods}
\alias{show_net}
\title{Show an Object}
\usage{
show_net(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} class}
}
\description{
show method short for PFPRefnet object, see \code{\link[methods]{show_net}}
}
\seealso{
\code{\link{PFPRefnet-class}}
}
