% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{show_PFP}
\alias{show_PFP}
\alias{show_PFP-methods}
\alias{show_PFP,PFP-method}
\title{The show_PFP generic function}
\usage{
show_PFP(object)

\S4method{show_PFP}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} object}
}
\description{
Show a short summary for PFP object, see \code{\link[methods]{show_PFP}}.
}
