% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{sub_PFP-methods}
\alias{sub_PFP-methods}
\alias{sub_PFP}
\alias{sub_PFP,PFP-method}
\title{subset of PFP object}
\usage{
sub_PFP(
  object,
  group_name = NULL,
  index = NULL,
  index_type = c("slice", "pathway_id", "pathway_name")
)

\S4method{sub_PFP}{PFP}(
  object,
  group_name = NULL,
  index = NULL,
  index_type = c("slice", "pathway_id", "pathway_name")
)
}
\arguments{
\item{object, }{\code{PFP} class}

\item{group_name, }{the group name in kegg}

\item{index, }{the index of pathway, NULL or a list contains slice/numeric, character, specifying elements to extract.
This parameter' length must be the same as \code{group_name}.
Default is \emph{NULL}, indicating extract all the networks of a group. See
\emph{details} for more information.}

\item{index_type, }{the index type,such as "slice","pathway_id","pathway_name"}
}
\value{
a PFP object contains just the selected elements.
}
\description{
This function extract the subsets of PFP-class.
}
\details{
This function help users to extract the specific networks PFPscores for
customized analysis, which could be of entire group PFP or some part of
a specific group PFP.

Note, the \code{index} argument is only worked while the group_name argument is
consideration, which means group_name is not \emph{NULL}. And the length must be
the same as \code{group_name}. Default is \emph{NULL}, indicating extract the entire PFP.
}
\examples{
# New a PFP object
data(PFP_test1)
#s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
#refnet_names <- sub_PFP(PFP_test1,group_name,index,index_type)
}
\seealso{
\code{\link{PFP-class}}
}
