% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{subnet-methods}
\alias{subnet-methods}
\alias{subnet}
\alias{subnet,PFPRefnet-method}
\title{Subset the basic networks}
\usage{
subnet(
  object,
  group_name = NULL,
  index = NULL,
  index_type = c("slice", "pathway_id", "pathway_name")
)

\S4method{subnet}{PFPRefnet}(
  object,
  group_name = NULL,
  index = NULL,
  index_type = c("slice", "pathway_id", "pathway_name")
)
}
\arguments{
\item{object, }{\code{PFPRefnet} class.}

\item{group_name, }{character, indicating the groups to subset.}

\item{index, }{NULL or a list contains slice/numeric, character, specifying elements to extract.
This parameter' length must be the same as \code{group_name}.
Default is \emph{NULL}, indicating extract all the networks of a group. See
\emph{details} for more information.}

\item{index_type, }{character, the type pf index, which could be "slice","id","name".}
}
\description{
Extract or Replace parts of the PFPRefnet.
}
\details{
This function help users to extract the specific networks for
customized analysis, which could be of entire group networks or some part of
a specific group networks.

Note, the \code{index} argument is only worked while the group_name argument is
consideration, which means group_name is not \emph{NULL}. And the length must be
the same as \code{group_name}. Default is \emph{NULL}, indicating extract the entire group basic networks.
}
\seealso{
\code{\link{PFPRefnet-class}}
}
