% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kegg_graphs.R
\name{trans_graph2PFPRefnet}
\alias{trans_graph2PFPRefnet}
\title{translate graph_list to PFPRefnet class}
\usage{
trans_graph2PFPRefnet(graph_list, pathway_info)
}
\arguments{
\item{graph_list, }{a list of \code{\link{graphNEL}}.}

\item{pathway_info, }{a data.frame, which contains all kegg pathways "index","id","name","group","species"}
}
\description{
This function will translate all graphs in \code{graph_list} to a \code{\link{PFPRefnet-class}} object.
}
\details{
, translating all graphs in \code{graph_list} to a \code{\link{PFPRefnet-class}} object.
The pathway_info can be designed by yourself, but the colnames must be "index","id","name","group" and "species".
}
\examples{
\dontrun{
data(pathway_info)
kegg_download(spec = "hsa", file_root="~/Desktop")
graph_list <- trans_xml2graph("~/Desktop")
PFPRefnet_hsa <- trans_graph2PFPRefnet(graph_list,pathway_info)
}
}
