% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{trans_id}
\alias{trans_id}
\title{translate other id type to ENTREZID}
\usage{
trans_id(gene_list, gene_id_type, gene_info_db, drop = FALSE)
}
\arguments{
\item{gene_list, }{a vector of characters}

\item{gene_id_type, }{a character,the type of gene ID, "ENSEMBL","GO","SYMBOL" and so on.}

\item{drop, }{a logical,whether to drop NA when translating id type}
}
\description{
translate other id type to ENTREZID
}
\details{
, translate other id type to ENTREZID,
}
\examples{
\dontrun{
library(org.Mm.eg.db)
data(gene_list)
gene_n <- trans_id(gene_list,gene_id_type)
}
}
