% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{gene_list_hsa}
\alias{gene_list_hsa}
\title{A gene list of human
gene_list_hsa is a array of 40 genetic ENTREZID.}
\format{
Gene list obtained by differential gene analysis
}
\description{
A gene list of human
gene_list_hsa is a array of 40 genetic ENTREZID.
}
\examples{
data(gene_list_hsa)
}
