% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{genes_score-methods}
\alias{genes_score-methods}
\alias{genes_score}
\alias{genes_score,PFP-method}
\title{The score of genes in \emph{PFP} class}
\usage{
genes_score(
  object,
  index = NULL,
  index_type = c("pathway_id", "pathway_name", "slice")
)

\S4method{genes_score}{PFP}(
  object,
  index = NULL,
  index_type = c("pathway_id", "pathway_name", "slice")
)
}
\arguments{
\item{object, }{\code{PFP} class}

\item{index, }{character, indicating the groups to subset.}

\item{index_type, }{"pathway_id","pathway_name","slice"}
}
\value{
a named vector of numeric scores
}
\description{
This function extract the detail scores of every gene in the gene_list by
specific condition.
}
\examples{
data(PFP_test1)
genes_score <- genes_score(PFP_test1)
}
\seealso{
\code{\link{PFP-class}}
}
