% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kegg_graphs.R
\name{kegg_download}
\alias{kegg_download}
\title{download kegg KGML files}
\usage{
kegg_download(spec, file_root = ".", test_mode = FALSE)
}
\arguments{
\item{spec, }{a character, refers to the species names in kegg, such as "hsa",
"mmu"...}

\item{file_root, }{a character,refers to the root you want to save kegg
pathway kgml files in.}

\item{test_mode, }{please set whether to test this function.}
}
\value{
the kegg KGML files
}
\description{
This function will download all kegg KGML files assigned by
\code{spec}.
}
\details{
Downloading all kegg KGML files assigned by \code{spec} from
https://www.kegg.jp/kegg/xml/,
which may take tens of minutes.
}
\examples{
kegg_download(spec,file_root=".", test_mode=TRUE)
}
