% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{pathways_score-methods}
\alias{pathways_score-methods}
\alias{pathways_score}
\alias{pathways_score-method}
\alias{pathways_score,PFP-method}
\title{Basic pathway networks scores of \emph{PFP} class
This function can extract the details in pathway fingerprint scores.}
\usage{
pathways_score(object)

\S4method{pathways_score}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} class}
}
\value{
as list, details in pathway fingerprint scores.
}
\description{
Basic pathway networks scores of \emph{PFP} class
This function can extract the details in pathway fingerprint scores.
}
\examples{
# New a PFP object
data(PFP_test1)
#s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
pathways_score <- pathways_score(PFP_test1)
}
\seealso{
\code{\link{PFP-class}}
}
