% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\name{PFPRefnet_mmu}
\alias{PFPRefnet_mmu}
\title{Pathway fingerprint data of mouse}
\format{
A PFPRefnet object, more details see \code{\link{PFPRefnet-class}}
}
\description{
A dataset containing the pathway maps of KEGG
}
\examples{
# Load the data.
data(PFPRefnet_mmu)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
