% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PFPlist.R
\name{plot_PFPlist}
\alias{plot_PFPlist}
\title{Plot multiple PFPs.}
\usage{
plot_PFPlist(object, l_size = 0.5)
}
\arguments{
\item{object, }{\code{PFP} a list of PFP.}

\item{l_size, }{line size of plot, default is 0.5.}
}
\value{
plot the PFP list
}
\description{
Function for visualization multiple PFPs.
}
\examples{
data(PFP_test1)
pfp_list <- list(a=PFP_test1)
plot_PFPlist(pfp_list)
}
\seealso{
\code{\link{PFP-class}}
}
