% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{rank_PFP-methods}
\alias{rank_PFP-methods}
\alias{rank_PFP}
\alias{rank_PFP,PFP-method}
\title{rank of the PFP object by the value of PFP_score.
Function for show the rank of PFP results.}
\usage{
rank_PFP(object, total_rank = FALSE, decreasing = TRUE)

\S4method{rank_PFP}{PFP}(object, total_rank = FALSE, decreasing = TRUE)
}
\arguments{
\item{object, }{\code{PFP} class}

\item{total_rank, }{a logical, whether to rank in total range, the default is \emph{TRUE}}

\item{decreasing, }{a logical, Sorting method, the default is \emph{TRUE}}
}
\value{
a ranked PFP object.
}
\description{
rank of the PFP object by the value of PFP_score.
Function for show the rank of PFP results.
}
\seealso{
\code{\link{PFP-class}}
}
