% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{rank_PFP-methods}
\alias{rank_PFP-methods}
\alias{rank_PFP}
\alias{rank_PFP,PFP-method}
\title{rank PFPscore}
\usage{
rank_PFP(
  object,
  total_rank = FALSE,
  decreasing = TRUE,
  thresh_slot = "p_adj_value",
  thresh_value = 0.05
)

\S4method{rank_PFP}{PFP}(
  object,
  total_rank = FALSE,
  decreasing = TRUE,
  thresh_slot = "p_adj_value",
  thresh_value = 0.05
)
}
\arguments{
\item{object, }{\code{PFP} class}

\item{total_rank, }{a logical, whether to rank in total range,the default is
\emph{TRUE}}

\item{decreasing, }{a logical, Sorting method, the default is \emph{TRUE}}

\item{thresh_slot, }{a character, it could be 'p_value' or 'p_adj_value',
it means the threshold slot to choose for select the significant pathway.
Default is 'p_adj_value'.It also could be \emph{NULL},it means that you don't
want to select the significant pathway and you will select all pathways.}

\item{thresh_value, }{a numeric, threshold value of 'p_value' or
'p_adjust_value' for pathway selection}
}
\value{
a ranked PFP object.
}
\description{
rank the PFP object by the value of PFP_score.
}
\examples{
# New a PFP object
data(PFP_test1)
rank_PFP(PFP_test1,
        total_rank=FALSE,
        decreasing=TRUE,
        thresh_slot="p_adj_value",
        thresh_value = 0.05)
}
\seealso{
\code{\link{PFP-class}}
}
