% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{refnet_info-methods}
\alias{refnet_info-methods}
\alias{refnet_info}
\alias{refnet_info,PFP-method}
\title{Basic network information of \emph{PFP} class}
\usage{
refnet_info(object)

\S4method{refnet_info}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} class}
}
\value{
detail information of reference pathway networks
}
\description{
This function extract the detail information of reference pathway networks.
}
\examples{
data(PFP_test1)
refnet_info <- refnet_info(PFP_test1)
}
\seealso{
\code{\link{PFP-class}}
}
