% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{refnet_names-methods}
\alias{refnet_names-methods}
\alias{refnet_names}
\alias{refnet_names,PFP-method}
\title{Names of basic networks
This function extract the reference pathway network names of PFP.}
\usage{
refnet_names(object)

\S4method{refnet_names}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} class}
}
\value{
a vector contains pathway names
}
\description{
Names of basic networks
This function extract the reference pathway network names of PFP.
}
\examples{
# New a PFP object
data(PFP_test1)
#s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
refnet_names <- refnet_names(PFP_test1)
}
