% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{show_net}
\alias{show_net}
\alias{show_net-methods}
\alias{show_net,PFPRefnet-method}
\title{Show an Object}
\usage{
show_net(object)

\S4method{show_net}{PFPRefnet}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} object}
}
\value{
show the network
}
\description{
show method short for PFPRefnet object
}
\examples{
# New a PFPRefnet object
s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
rank_PFP(s1,'line', total_rank=FALSE,decreasing=TRUE,separate=TRUE,p_adj = 0.05)
# New a PFPRefnet object
PFPRefnet1 <- new("PFPRefnet", network = list1, net_info = data1)
show_net(PFPRefnet1)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
