% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFPRefnet-class.R
\docType{methods}
\name{show_net}
\alias{show_net}
\alias{show_net-methods}
\alias{show_net,PFPRefnet-method}
\title{Show an Object}
\usage{
show_net(object)

\S4method{show_net}{PFPRefnet}(object)
}
\arguments{
\item{object, }{\code{PFPRefnet} object}
}
\value{
show the network
}
\description{
show method short for PFPRefnet object
}
\examples{
data(PFPRefnet_hsa)
show_net(PFPRefnet_hsa)
}
\seealso{
\code{\link{PFPRefnet-class}}
}
