% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFP-class.R
\docType{methods}
\name{show_PFP}
\alias{show_PFP}
\alias{show_PFP-methods}
\alias{show_PFP,PFP-method}
\title{The show_PFP generic function}
\usage{
show_PFP(object)

\S4method{show_PFP}{PFP}(object)
}
\arguments{
\item{object, }{\code{PFP} object}
}
\value{
show the PFP
}
\description{
Show a short summary for PFP object.
}
\examples{
# New a PFP object
data(PFP_test1)
#s1 <- new("PFP", pathways_score = list1, refnet_info = data1)
show_PFP(PFP_test1)
}
