% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kegg_graphs.R
\name{trans_graph2PFPRefnet}
\alias{trans_graph2PFPRefnet}
\title{translate graph_list to PFPRefnet class}
\usage{
trans_graph2PFPRefnet(graph_list, pathway_info)
}
\arguments{
\item{graph_list, }{a list of \code{\link{graphNEL}}.}

\item{pathway_info, }{a data.frame, which contains all kegg pathways "index",
"id","name","group","species"}
}
\value{
a PFPRefnet
}
\description{
This function will translate all graphs in \code{graph_list} to
a \code{\link{PFPRefnet-class}} object.
}
\details{
translating all graphs in \code{graph_list} to a
\code{\link{PFPRefnet-class}} object.
The pathway_info can be designed by yourself, but the colnames must be
"index","id","name","group" and "species".
}
\examples{
# Load the info of the pathway
data(pathway_info)
pathway_info
}
