% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kegg_graphs.R
\name{trans_xml2graph}
\alias{trans_xml2graph}
\title{translate kgml files to graphNEl}
\usage{
trans_xml2graph(file_dir)
}
\arguments{
\item{file_dir, }{a character, refers to the file_path where kegg KGML files are stored.}
}
\description{
This function will translate all kegg KGML files in path \code{file_dir}.
}
\details{
, translate all kegg KGML files in path \code{file_dir}. It will return a list of \code{graphNEL}
}
\examples{
\dontrun{
kegg_download(spec = "hsa", file_root="~/Desktop")
graph_list <- trans_xml2graph("~/Desktop")
}
}
